Restructure several places
Added DeploymentManager as a thread for managing the deployed service
including steps for deployment, undeployment, chaining, monitoring.
Now the RestInterfaceServer sends a message to the DeploymentManager
for deployment.
The DeploymentManager calls the PlacementPlugin and translation of the
service and deploys the stacks and reacts on further messages such as
updates/ tear down.
MessageQueue now has only one queue - the queue for the DeploymentManager.
The TranslatorCore now starts the RestInterfaceServer and the DeploymentManager.
It waits for the DeploymentManager to die to prevent application stop.
On shutdown of the application a terminate message is sent to the
DeploymentManager to clean up the stacks - this already works.
Some code moved from the DescriptorTranslator to the DeploymentManager.