Add new dependency + fix logging issue
The new dependency org.apache.httpcomponents.httpasyncclient
uses apache logging which uses an old version of a lib.
This results in an interface conflict for the logging system.
By also adding the new(!) version of org.slf4j.jcl-over-slf4j
the can be fixed.
For debugging reasons several exclusions were added that
might not be necessary.